put subparameters into nested arguments
For optional input port / parameters, the reference in implementation/args must be placed in
a nested list to indicate that it is optional.

Refer to https://aka.ms/azureml-module-optional-inputs for details.

For the subparameters (the ones that under another parameter's `options` section),
should also put the related args into nested lists, otherwise the module will fail to register
in the latest CLI tool.

This PR is to update the module spec to fix the issue.